requirements.txt
jax[cuda12_pip]>=0.4.20
jaxlib>=0.4.20
numpy>=1.24.0
cupy-cuda12x>=12.0.0
pyyaml>=6.0
joblib>=1.3.0
open-spiel>=1.3
scipy>=1.10.0
matplotlib>=3.6.0
tqdm>=4.65.0
tensorboard>=2.13.0
pytest>=7.4.0